refactor(mdx): ♻️ move serialized content contract into entities
Move SerializedMDXContent to entities and update entity models to import it there.
Update MDX serializer and resolver to use the shared contract.

Made-with: Cursor
