+++ src/entities/mdx-content/models/mdx-content.ts
@@ -1,0 +1,4 @@
+export type SerializedMDXContent = {
+  type: "serialized";
+  data: string;
+};
